UpdatePrayerCount Lambda Function
Updated lambda function to fix the errors I was getting back from the
API regarding incorrect schema. It turns out you need to send the
primary and sort key when updating a DynamoDB with a sort key. I also
created a new layer in Lambda for the imports. This isn't part of the
project source code.