Prayer Count, Props And Context API
The bulk of the work went into getting the prayer count to increase and
send the change to the database.

The Prayers-Context was updated to switch between different fetch types
and include parameters required for put. I'll need to refactor this
file as its a bit messy.

Card component setPrayer method was updated to extract the current
prayer count and relay the payload to the Prayers-Context. Also, in
order to get the prayer count I utilized `forwardedRef` in the
component.

The PrayerWall module Card component props were simplified to key and
data.