Recaptcha
Added `react-google-recaptcha` library to project to help me integrate
Recaptcha into the PostPrayer component. Now I know why Recaptcha looks
horrendous on websites because there's no way of styling it. Google in
their wisdom used an iframe and the v3 API doesn't have any way of
changing anything. Anyways, I updated the submission logic to check for
the Recaptcha to be checked. I still need to update it so the Recaptcha
token is sent to the Lambda function and then verified by Google's API.