Recaptcha Token And Instructions
Updated the PostPrayer form to include and send the Recaptcha token in
`formData`. At the moment, its not verifying with Google API but will
soon.

I also added a concise message to encourage users to provide their
prayer request while indicating that they can use commas and periods
for punctuation. I'll do the same for the input field error as I am not
happy with its current verbiage.