PrayerWall Array Check
Updated conditional to check if the response is an array. If its not
then output body response from the server. This takes care of the case
where there are no items in a database. Technically, not a query error
but more of I haven't put any prayers in the database.