Hero Component Image Refactor
Refactored the hero component to display the background image inline
rather than using css to display it. I did this to increase the
pagespeed score. The assumption here is I was getting penalized for the
delay of the image displaying because the css needs to load first.

Coding this was a bit trickier because I used Vite's `import.meta.glob`
to import all the hero images versus importing individually. In a small
project like this it probably would be ok but at scale it would be an
elegant solution.

I'm going to revise this component again to make it reusable.